[CI] Warm the TTS-stage ASR the same way as the standalone gate (#911)
