+++ sglang_omni/models/qwen3_asr/config.py
@@ -25,6 +25,7 @@ class Qwen3ASRPipelineConfig(PipelineConfig):
             factory_args={
                 "device": "cuda:0",
                 "max_running_requests": 32,
+                "max_new_tokens": 128,
                 "request_build_max_workers": 2,
                 "request_build_max_pending": 16,
             },
