+++ tests/utils.py
@@ -13,7 +13,7 @@ from typing import TYPE_CHECKING
 import pytest
 
 from benchmarks.benchmarker import utils as benchmark_utils
-from benchmarks.tasks.tts import (
+from benchmarks.tasks.asr import (
     DEFAULT_ASR_TRANSCRIBE_CONCURRENCY,
     QWEN3_ASR_MODEL_PATH,
 )
